requirements.txt
supabase==2.9.0
python-dotenv==1.0.0
Pillow==10.2.0
httpx==0.27.0
